refactor out client crypto module #130
